LayoutTests/inspector/model/remote-object/dom-expected.txt


-----------------------------------------------------
EXPRESSION: window.loadEvent
{
  "_type": "object",
  "_objectId": "<filtered>",
  "_description": "Event",
  "_preview": {
    "_type": "object",
    "_description": "Event",
    "_lossless": false,
    "_overflow": true,
    "_properties": [
      {
        "_name": "isTrusted",
        "_type": "boolean",
        "_value": "true"
      },
      {
        "_name": "type",
        "_type": "string",
        "_value": "load"
      },
      {
        "_name": "target",
        "_type": "object",
        "_subtype": "node",
        "_value": "#document"
      },
      {
        "_name": "currentTarget",
        "_type": "object",
        "_subtype": "null",
        "_value": "null"
      },
      {
        "_name": "eventPhase",
        "_type": "number",
        "_value": "0"
      }
    ],
    "_entries": null
  }
}

-----------------------------------------------------
EXPRESSION: document.body
{
  "_type": "object",
  "_subtype": "node",
  "_objectId": "<filtered>",
  "_description": "<body>",
  "_preview": {
    "_type": "object",
    "_subtype": "node",
    "_description": "<body>",
    "_lossless": false,
    "_overflow": true,
    "_properties": [
      {
        "_name": "aLink",
        "_type": "string",
        "_value": ""
      },
      {
        "_name": "background",
        "_type": "string",
        "_value": ""
      },
      {
        "_name": "bgColor",
        "_type": "string",
        "_value": ""
      },
      {
        "_name": "link",
        "_type": "string",
        "_value": ""
      },
      {
        "_name": "text",
        "_type": "string",
        "_value": ""
      }
    ],
    "_entries": null
  }
}

-----------------------------------------------------
EXPRESSION: div = document.createElement("div"); div.className = "foo bar"; div
{
  "_type": "object",
  "_subtype": "node",
  "_objectId": "<filtered>",
  "_description": "<div class=\"foo bar\">",
  "_preview": {
    "_type": "object",
    "_subtype": "node",
    "_description": "<div class=\"foo bar\">",
    "_lossless": false,
    "_overflow": true,
    "_properties": [
      {
        "_name": "align",
        "_type": "string",
        "_value": ""
      },
      {
        "_name": "title",
        "_type": "string",
        "_value": ""
      },
      {
        "_name": "lang",
        "_type": "string",
        "_value": ""
      },
      {
        "_name": "translate",
        "_type": "boolean",
        "_value": "true"
      },
      {
        "_name": "dir",
        "_type": "string",
        "_value": ""
      }
    ],
    "_entries": null
  }
}

-----------------------------------------------------
EXPRESSION: span = document.createElement("span"); span.id = "foo"; span
{
  "_type": "object",
  "_subtype": "node",
  "_objectId": "<filtered>",
  "_description": "<span id=\"foo\">",
  "_preview": {
    "_type": "object",
    "_subtype": "node",
    "_description": "<span id=\"foo\">",
    "_lossless": false,
    "_overflow": true,
    "_properties": [
      {
        "_name": "title",
        "_type": "string",
        "_value": ""
      },
      {
        "_name": "lang",
        "_type": "string",
        "_value": ""
      },
      {
        "_name": "translate",
        "_type": "boolean",
        "_value": "true"
      },
      {
        "_name": "dir",
        "_type": "string",
        "_value": ""
      },
      {
        "_name": "cue",
        "_type": "boolean",
        "_value": "false"
      }
    ],
    "_entries": null
  }
}

-----------------------------------------------------
EXPRESSION: input = document.createElement("input"); input.type = "password"; input
{
  "_type": "object",
  "_subtype": "node",
  "_objectId": "<filtered>",
  "_description": "<input type=\"password\">",
  "_preview": {
    "_type": "object",
    "_subtype": "node",
    "_description": "<input type=\"password\">",
    "_lossless": false,
    "_overflow": true,
    "_properties": [
      {
        "_name": "accept",
        "_type": "string",
        "_value": ""
      },
      {
        "_name": "alpha",
        "_type": "boolean",
        "_value": "false"
      },
      {
        "_name": "alt",
        "_type": "string",
        "_value": ""
      },
      {
        "_name": "autocomplete",
        "_type": "string",
        "_value": ""
      },
      {
        "_name": "defaultChecked",
        "_type": "boolean",
        "_value": "false"
      }
    ],
    "_entries": null
  }
}

-----------------------------------------------------
EXPRESSION: text = document.createTextNode("text content"); text
{
  "_type": "object",
  "_subtype": "node",
  "_objectId": "<filtered>",
  "_description": "#text \"text content\"",
  "_preview": {
    "_type": "object",
    "_subtype": "node",
    "_description": "#text \"text content\"",
    "_lossless": false,
    "_overflow": true,
    "_properties": [
      {
        "_name": "wholeText",
        "_type": "string",
        "_value": "text content"
      },
      {
        "_name": "assignedSlot",
        "_type": "object",
        "_subtype": "null",
        "_value": "null"
      },
      {
        "_name": "splitText",
        "_type": "function",
        "_value": ""
      },
      {
        "_name": "data",
        "_type": "string",
        "_value": "text content"
      },
      {
        "_name": "length",
        "_type": "number",
        "_value": "12"
      }
    ],
    "_entries": null
  }
}

-----------------------------------------------------
EXPRESSION: comment = document.createComment("comment content"); comment
{
  "_type": "object",
  "_subtype": "node",
  "_objectId": "<filtered>",
  "_description": "<!--comment content-->",
  "_preview": {
    "_type": "object",
    "_subtype": "node",
    "_description": "<!--comment content-->",
    "_lossless": false,
    "_overflow": true,
    "_properties": [
      {
        "_name": "data",
        "_type": "string",
        "_value": "comment content"
      },
      {
        "_name": "length",
        "_type": "number",
        "_value": "15"
      },
      {
        "_name": "previousElementSibling",
        "_type": "object",
        "_subtype": "null",
        "_value": "null"
      },
      {
        "_name": "nextElementSibling",
        "_type": "object",
        "_subtype": "null",
        "_value": "null"
      },
      {
        "_name": "substringData",
        "_type": "function",
        "_value": ""
      }
    ],
    "_entries": null
  }
}

-----------------------------------------------------
EXPRESSION: svgElement = document.createElementNS("http://www.w3.org/2000/svg", "rect"); svgElement.classList.add("test"); svgElement
{
  "_type": "object",
  "_subtype": "node",
  "_objectId": "<filtered>",
  "_description": "<rect class=\"test\">",
  "_preview": {
    "_type": "object",
    "_subtype": "node",
    "_description": "<rect class=\"test\">",
    "_lossless": false,
    "_overflow": true,
    "_properties": [
      {
        "_name": "x",
        "_type": "object",
        "_value": "SVGAnimatedLength"
      },
      {
        "_name": "y",
        "_type": "object",
        "_value": "SVGAnimatedLength"
      },
      {
        "_name": "width",
        "_type": "object",
        "_value": "SVGAnimatedLength"
      },
      {
        "_name": "height",
        "_type": "object",
        "_value": "SVGAnimatedLength"
      },
      {
        "_name": "rx",
        "_type": "object",
        "_value": "SVGAnimatedLength"
      }
    ],
    "_entries": null
  }
}

-----------------------------------------------------
EXPRESSION: [document.body, div, span, input, text, comment, svgElement]
{
  "_type": "object",
  "_subtype": "array",
  "_objectId": "<filtered>",
  "_description": "Array",
  "_size": 7,
  "_preview": {
    "_type": "object",
    "_subtype": "array",
    "_description": "Array",
    "_lossless": false,
    "_overflow": false,
    "_size": 7,
    "_properties": [
      {
        "_name": "0",
        "_type": "object",
        "_subtype": "node",
        "_value": "<body>"
      },
      {
        "_name": "1",
        "_type": "object",
        "_subtype": "node",
        "_value": "<div class=\"foo bar\">"
      },
      {
        "_name": "2",
        "_type": "object",
        "_subtype": "node",
        "_value": "<span id=\"foo\">"
      },
      {
        "_name": "3",
        "_type": "object",
        "_subtype": "node",
        "_value": "<input type=\"password\">"
      },
      {
        "_name": "4",
        "_type": "object",
        "_subtype": "node",
        "_value": "#text \"text content\""
      },
      {
        "_name": "5",
        "_type": "object",
        "_subtype": "node",
        "_value": "<!--comment content-->"
      },
      {
        "_name": "6",
        "_type": "object",
        "_subtype": "node",
        "_value": "<rect class=\"test\">"
      }
    ],
    "_entries": null
  }
}

-----------------------------------------------------
EXPRESSION: document.head.children
{
  "_type": "object",
  "_subtype": "array",
  "_objectId": "<filtered>",
  "_description": "HTMLCollection",
  "_size": 4,
  "_preview": {
    "_type": "object",
    "_subtype": "array",
    "_description": "HTMLCollection",
    "_lossless": false,
    "_overflow": false,
    "_size": 4,
    "_properties": [
      {
        "_name": "0",
        "_type": "object",
        "_subtype": "node",
        "_value": "<meta>"
      },
      {
        "_name": "1",
        "_type": "object",
        "_subtype": "node",
        "_value": "<script>"
      },
      {
        "_name": "2",
        "_type": "object",
        "_subtype": "node",
        "_value": "<script>"
      },
      {
        "_name": "3",
        "_type": "object",
        "_subtype": "node",
        "_value": "<script>"
      }
    ],
    "_entries": null
  }
}

-----------------------------------------------------
EXPRESSION: document.getElementsByClassName("my-test")
{
  "_type": "object",
  "_subtype": "array",
  "_objectId": "<filtered>",
  "_description": "HTMLCollection",
  "_size": 3,
  "_preview": {
    "_type": "object",
    "_subtype": "array",
    "_description": "HTMLCollection",
    "_lossless": false,
    "_overflow": false,
    "_size": 3,
    "_properties": [
      {
        "_name": "0",
        "_type": "object",
        "_subtype": "node",
        "_value": "<p class=\"my-test\">"
      },
      {
        "_name": "1",
        "_type": "object",
        "_subtype": "node",
        "_value": "<p class=\"my-test\">"
      },
      {
        "_name": "2",
        "_type": "object",
        "_subtype": "node",
        "_value": "<p class=\"my-test\">"
      }
    ],
    "_entries": null
  }
}

-----------------------------------------------------
EXPRESSION: document.querySelectorAll(".my-test")
{
  "_type": "object",
  "_subtype": "array",
  "_objectId": "<filtered>",
  "_description": "NodeList",
  "_size": 3,
  "_preview": {
    "_type": "object",
    "_subtype": "array",
    "_description": "NodeList",
    "_lossless": false,
    "_overflow": false,
    "_size": 3,
    "_properties": [
      {
        "_name": "0",
        "_type": "object",
        "_subtype": "node",
        "_value": "<p class=\"my-test\">"
      },
      {
        "_name": "1",
        "_type": "object",
        "_subtype": "node",
        "_value": "<p class=\"my-test\">"
      },
      {
        "_name": "2",
        "_type": "object",
        "_subtype": "node",
        "_value": "<p class=\"my-test\">"
      }
    ],
    "_entries": null
  }
}
